feat: improve UTF string handling and size_t compatibility
* Enhance UTF string conversion and buffer handling
  - Improve buffer conversion efficiency by batching character adds
  - Add mulle_utf8data introspection capabilities
  - Optimize UTF32 to UTF8 conversion paths

* Strengthen type safety and consistency
  - Replace unsigned int with size_t for string lengths
  - Change mulle_utf8_t to char for better C compatibility
  - Update function signatures to use size_t consistently

* Breaking changes
  - Rename mulle_unicode_is_privatecharacterplane to mulle_utf_is_privatecharacterplane
  - Change parameter order in _strcnpy functions
  - Update mulle_utf_add_bytes_function_t typedef
  - Move stdlib-like code to mulle-utf8-string.c

* Other improvements
  - Increase test coverage significantly
  - Fix 32-bit compatibility issues
  - Update CMake configuration for better Windows support

+++ src/mulle-utf8.h
@@ -93,7 +93,7 @@ static inline int   mulle_utf8_is_validcontinuationcharacter( char c)
 
 
 // length excluding 'c'
-static inline unsigned int  mulle_utf8_get_extracharacterslength( char c)
+static inline size_t  mulle_utf8_get_extracharacterslength( char c)
 {
    assert( mulle_utf8_get_startcharactertype( c) == mulle_utf8_multiple_start_character);
 
@@ -120,10 +120,10 @@ static inline unsigned int  mulle_utf8_get_extracharacterslength( char c)
 // returned length does not include BOM
 //
 MULLE__UTF_GLOBAL
-unsigned int  mulle_utf8_utf16length( char *src, size_t len);
+size_t  mulle_utf8_utf16length( char *src, size_t len);
 
 MULLE__UTF_GLOBAL
-unsigned int  mulle_utf8_utf32length( char *src, size_t len);
+size_t  mulle_utf8_utf32length( char *src, size_t len);
 
 
 static inline size_t  mulle_utf8_utf16maxlength( size_t len)
